+++ crates/o5/src/handshake_old/integration.rs
@@ -15,175 +15,248 @@
 //! <!-- @@ end test lint list maintained by maint/add_warning @@ -->
 
 use super::*;
-use crate::common::ntor_arti::{ClientHandshake, ServerHandshake};
+use crate::{
+    common::{colorize, curve25519},
+    test_utils::{init_subscriber, FakePRNG},
+};
 
 use hex_literal::hex;
 use tor_basic_utils::test_rng::testing_rng;
-use digest::XofReader;
 
-
+fn make_fake_ephem_key(bytes: &[u8]) -> curve25519::EphemeralSecret {
+    assert_eq!(bytes.len(), 32);
+    let rng = FakePRNG::new(bytes);
+    curve25519::EphemeralSecret::random_from_rng(rng)
+}
 
 #[test]
-fn test_obfs4_roundtrip() {
+fn test_o5_roundtrip() -> Result<()> {
     let mut rng = rand::thread_rng();
-    let relay_private = O5NtorSecretKey::generate_for_test(&mut testing_rng());
 
-    let verification = &b"shared secret"[..];
-    let client_message = &b"Hello. I am a client. Let's be friends!"[..];
-    let relay_message = &b"Greetings, client. I am a robot. Beep boop."[..];
+    let relay_private = O5NtorSecretKey::generate_for_test(&mut rng);
+    let x = O5NtorSecretKey::generate_for_test(&mut rng);
+    let y = O5NtorSecretKey::generate_for_test(&mut rng);
 
-    let (c_state, c_handshake) =
-        client_handshake_obfs4(&mut rng, &relay_private.pk, client_message, verification)
-            .unwrap();
+    let mut sid = [0u8; SESSION_ID_LEN];
+    rand::thread_rng().fill_bytes(&mut sid);
 
-    struct Rep(Vec<u8>, Vec<u8>);
-    impl MsgReply for Rep {
-        fn reply(&mut self, msg: &[u8]) -> Option<Vec<u8>> {
-            self.0 = msg.to_vec();
-            Some(self.1.clone())
-        }
-    }
-    let mut rep = Rep(Vec::new(), relay_message.to_vec());
+    let chs_materials = CHSMaterials::new(relay_private.pk, colorize(sid));
 
-    let (s_handshake, mut s_keygen) = server_handshake_obfs4(
-        &mut rng,
-        &mut rep,
-        &c_handshake,
-        &[relay_private],
-        verification,
-    )
-    .unwrap();
+    let server = Server::new_from_key(relay_private);
 
-    let (s_msg, mut c_keygen) =
-        client_handshake_obfs4_part2(&c_state, &s_handshake, verification).unwrap();
+    let shs_materials = SHSMaterials {
+        identity_keys: server.identity_keys.clone(),
+        len_seed: [0u8; SEED_LENGTH],
+        session_id: "s-yyy".into(),
+    };
 
-    assert_eq!(rep.0[..], client_message[..]);
-    assert_eq!(s_msg[..], relay_message[..]);
-    let mut s_keys = [0_u8; 100];
-    let mut c_keys = [0_u8; 1000];
-    s_keygen.read(&mut s_keys);
-    c_keygen.read(&mut c_keys);
-    assert_eq!(s_keys[..], c_keys[..100]);
+    let (state, create_msg) = client_handshake_o5_no_keygen(x.sk, chs_materials).unwrap();
+
+    let ephem = make_fake_ephem_key(&y.sk.as_bytes()[..]);
+    let (s_keygen, created_msg) = server
+        .server_handshake_o5_no_keygen(ephem, &create_msg[..], shs_materials)
+        .unwrap();
+
+    let (c_keygen, _) = client_handshake2_o5(created_msg, &state)?;
+
+    let c_keys = c_keygen.expand(72)?;
+    let s_keys = s_keygen.expand(72)?;
+    assert_eq!(c_keys, s_keys);
+
+    Ok(())
 }
 
 // Same as previous test, but use the higher-level APIs instead.
 #[test]
-fn test_obfs4_roundtrip_highlevel() {
-    let mut rng = rand::thread_rng();
-    let relay_private = O5NtorSecretKey::generate_for_test(&mut testing_rng());
-
-    let (c_state, c_handshake) =
-        Obfs4NtorClient::client1(&mut rng, &relay_private.pk, &[]).unwrap();
-
-    let mut rep = |_: &[NtorV3Extension]| Some(vec![]);
-
-    let (s_keygen, s_handshake) =
-        Obfs4NtorServer::server(&mut rng, &mut rep, &[relay_private], &c_handshake).unwrap();
-
-    let (extensions, keygen) = Obfs4NtorClient::client2(c_state, s_handshake).unwrap();
-
-    assert!(extensions.is_empty());
-    let c_keys = keygen.expand(1000).unwrap();
-    let s_keys = s_keygen.expand(100).unwrap();
-    assert_eq!(s_keys[..], c_keys[..100]);
-}
-
-// Same as previous test, but encode some congestion control extensions.
-#[test]
-fn test_obfs4_roundtrip_highlevel_cc() {
-    let mut rng = rand::thread_rng();
-    let relay_private = O5NtorSecretKey::generate_for_test(&mut testing_rng());
-
-    let client_exts = vec![NtorV3Extension::RequestCongestionControl];
-    let reply_exts = vec![NtorV3Extension::AckCongestionControl { sendme_inc: 42 }];
-
-    let (c_state, c_handshake) = Obfs4NtorClient::client1(
-        &mut rng,
-        &relay_private.pk,
-        &[NtorV3Extension::RequestCongestionControl],
-    )
-    .unwrap();
-
-    let mut rep = |msg: &[NtorV3Extension]| -> Option<Vec<NtorV3Extension>> {
-        assert_eq!(msg, client_exts);
-        Some(reply_exts.clone())
+fn test_o5_roundtrip_highlevel() -> Result<()> {
+    let rng = testing_rng();
+    let relay_secret = StaticSecret::random_from_rng(rng);
+    let relay_public = PublicKey::from(&relay_secret);
+    let relay_identity = RsaIdentity::from_bytes(&[12; 20]).unwrap();
+    let relay_ntpk = O5NtorPublicKey {
+        id: relay_identity,
+        pk: relay_public,
     };
+    let hs_materials = CHSMaterials::new(relay_ntpk, "c-xxx".into());
+    let (state, cmsg) = O5NtorHandshake::client1(&hs_materials, &())?;
 
-    let (s_keygen, s_handshake) =
-        Obfs4NtorServer::server(&mut rng, &mut rep, &[relay_private], &c_handshake).unwrap();
+    let relay_ntsk = O5NtorSecretKey {
+        pk: relay_ntpk,
+        sk: relay_secret,
+    };
+    let server = Server::new_from_key(relay_ntsk.clone());
+    let shs_materials = [SHSMaterials::new(
+        &relay_ntsk,
+        "s-yyy".into(),
+        [0u8; SEED_LENGTH],
+    )];
 
-    let (extensions, keygen) = Obfs4NtorClient::client2(c_state, s_handshake).unwrap();
+    let (skeygen, smsg) = server
+        .server(&mut |_: &()| Some(()), &shs_materials, &cmsg)
+        .unwrap();
 
-    assert_eq!(extensions, reply_exts);
-    let c_keys = keygen.expand(1000).unwrap();
-    let s_keys = s_keygen.expand(100).unwrap();
-    assert_eq!(s_keys[..], c_keys[..100]);
+    let (_extensions, ckeygen) = O5NtorHandshake::client2(state, smsg)?;
+
+    let skeys = skeygen.expand(55)?;
+    let ckeys = ckeygen.expand(55)?;
+
+    assert_eq!(skeys, ckeys);
+
+    Ok(())
 }
 
 #[test]
-fn test_obfs4_testvec() {
-    let id = hex!("9fad2af287ef942632833d21f946c6260c33fae6172b60006e86e4a6911753a2");
-    let b = hex!("a8649010896d3c05ab0e2ab75e3e9368695c8f72652e8aa73016756007054e59"); // identity key
-    let x = hex!("f886d140047a115b228a8f7f63cbc5a3ad74e9c970ad3cb4a7f7fd8067f0dd68"); // client session key
-    let y = hex!("403ec0927a8852bdff12129244fdb03cb3c94b8b8d15c863462fcda52b510d73"); // server session key
-    let b: curve25519::StaticSecret = b.into();
-    let B: curve25519::PublicKey = (&b).into();
-    let id: Ed25519Identity = id.into();
-    let x: curve25519::StaticSecret = x.into();
-    let y: curve25519::StaticSecret = y.into();
+fn test_o5_testvec_compat() -> Result<()> {
+    init_subscriber();
+    let b_sk = hex!("a83fdd04eb9ed77a2b38d86092a09a1cecfb93a7bdec0da35e542775b2e7af6e");
+    let x_sk = hex!("308ff4f3a0ebe8c1a93bcd40d67e3eec6b856aa5c07ef6d5a3d3cedf13dcf150");
+    let y_sk = hex!("881f9ad60e0833a627f0c47f5aafbdcb0b5471800eaeaa1e678291b947e4295c");
+    let id = hex!("000102030405060708090a0b0c0d0e0f10111213");
+    let expected_seed = "05b858d18df21a01566c74d39a5b091b4415f103c05851e77e79b274132dc5b5";
+    let expected_auth = "dc71f8ded2e56f829f1b944c1e94357fa8b7987f10211a017e2d1f2455092917";
 
-    let client_message = hex!("68656c6c6f20776f726c64");
-    let verification = hex!("78797a7a79");
-    let server_message = hex!("486f6c61204d756e646f");
-
-    let identity_public = O5NtorPublicKey { pk: B, id, rp: None };
-    let identity_private = O5NtorSecretKey {
-        sk: b,
-        pk: identity_public.clone(),
+    let sk: StaticSecret = b_sk.into();
+    let pk = O5NtorPublicKey {
+        id: RsaIdentity::from_bytes(&id).unwrap(),
+        pk: (&sk).into(),
     };
+    let relay_sk = O5NtorSecretKey { pk, sk };
+    let server = Server::new_from_key(relay_sk.clone());
 
-    let (state, client_handshake) =
-        client_handshake_obfs4_no_keygen(&identity_public, &client_message, &verification, x)
-            .unwrap();
+    let x: StaticSecret = x_sk.into();
 
-    // assert_eq!(client_handshake[..], hex!("9fad2af287ef942632833d21f946c6260c33fae6172b60006e86e4a6911753a2f8307a2bc1870b00b828bb74dbb8fd88e632a6375ab3bcd1ae706aaa8b6cdd1d252fe9ae91264c91d4ecb8501f79d0387e34ad8ca0f7c995184f7d11d5da4f463bebd9151fd3b47c180abc9e044d53565f04d82bbb3bebed3d06cea65db8be9c72b68cd461942088502f67")[..]);
+    let chs_materials = CHSMaterials::new(pk, "c-xxx".into());
 
-    struct Replier(Vec<u8>, Vec<u8>, bool);
-    impl MsgReply for Replier {
-        fn reply(&mut self, msg: &[u8]) -> Option<Vec<u8>> {
-            assert_eq!(msg, &self.0);
-            self.2 = true;
-            Some(self.1.clone())
-        }
-    }
-    let mut rep = Replier(client_message.to_vec(), server_message.to_vec(), false);
+    let shs_materials =
+        SHSMaterials::new(&server.identity_keys, "s-yyy".into(), [0u8; SEED_LENGTH]);
 
-    let (server_handshake, mut server_keygen) = server_handshake_obfs4_no_keygen(
-        &mut rep,
-        &y,
-        &client_handshake,
-        &[identity_private],
-        &verification,
+    let (state, create_msg) = client_handshake_o5_no_keygen(x, chs_materials).unwrap();
+
+    let (s_keygen, created_msg) = server
+        .server_handshake_o5_no_keygen(
+            make_fake_ephem_key(&y_sk[..]), // convert the StaticSecret to an EphemeralSecret for api to allow from hex
+            &create_msg[..],
+            shs_materials,
+        )
+        .unwrap();
+
+    let (c_keygen, auth) = client_handshake2_no_auth_check_o5(created_msg, &state)?;
+    let seed = c_keygen.seed.clone();
+
+    let c_keys = c_keygen.expand(72)?;
+    let s_keys = s_keygen.expand(72)?;
+
+    assert_eq!(&s_keys[..], &c_keys[..]);
+    assert_eq!(hex::encode(auth), expected_auth);
+    assert_eq!(hex::encode(&seed[..]), expected_seed);
+
+    Ok(())
+}
+
+#[cfg(target_feature = "disabled")]
+#[test]
+fn test_ntor_v1_testvec() -> Result<()> {
+    let b_sk = hex!("4820544f4c4420594f5520444f474954204b454550532048415050454e494e47");
+    let x_sk = hex!("706f6461792069207075742e2e2e2e2e2e2e2e4a454c4c59206f6e2074686973");
+    let y_sk = hex!("70686520737175697272656c2e2e2e2e2e2e2e2e686173206869732067616d65");
+    let id = hex!("69546f6c64596f7541626f75745374616972732e");
+    let client_handshake = hex!("69546f6c64596f7541626f75745374616972732eccbc8541904d18af08753eae967874749e6149f873de937f57f8fd903a21c471e65dfdbef8b2635837fe2cebc086a8096eae3213e6830dc407516083d412b078");
+    let server_handshake = hex!("390480a14362761d6aec1fea840f6e9e928fb2adb7b25c670be1045e35133a371cbdf68b89923e1f85e8e18ee6e805ea333fe4849c790ffd2670bd80fec95cc8");
+    let keys = hex!("0c62dee7f48893370d0ef896758d35729867beef1a5121df80e00f79ed349af39b51cae125719182f19d932a667dae1afbf2e336e6910e7822223e763afad0a13342157969dc6b79");
+
+    let sk: StaticSecret = b_sk.into();
+    let pk = O5NtorPublicKey {
+        id: RsaIdentity::from_bytes(&id).unwrap(),
+        pk: (&sk).into(),
+        rp: (&sk).into(),
+    };
+    let relay_sk = O5NtorSecretKey { pk, sk };
+
+    let x: StaticSecret = x_sk.into();
+    let y: StaticSecret = y_sk.into();
+
+    let (state, create_msg) =
+        client_handshake_o5_no_keygen((&x).into(), x, &relay_sk.pk).unwrap();
+    assert_eq!(&create_msg[..], &client_handshake[..]);
+
+    let (s_keygen, created_msg) = server_handshake_o5_no_keygen(
+        (&y).into(),
+        make_fake_ephem_key(&y_sk[..]), // convert the StaticSecret to an EphemeralSecret for api to allow from hex
+        &create_msg[..],
+        &[relay_sk],
     )
     .unwrap();
-    assert!(rep.2);
 
-    // assert_eq!(server_handshake[..], hex!("4bf4814326fdab45ad5184f5518bd7fae25dc59374062698201a50a22954246d2fc5f8773ca824542bc6cf6f57c7c29bbf4e5476461ab130c5b18ab0a91276651202c3e1e87c0d32054c")[..]);
+    assert_eq!(&created_msg[..], &server_handshake[..]);
 
-    let (server_msg_received, mut client_keygen) =
-        client_handshake_obfs4_part2(&state, &server_handshake, &verification).unwrap();
-    assert_eq!(&server_msg_received, &server_message);
+    let c_keygen = client_handshake2_o5(created_msg, &state)?;
 
-    let (c_keys, s_keys) = {
-        let mut c = [0_u8; 256];
-        let mut s = [0_u8; 256];
-        client_keygen.read(&mut c);
-        server_keygen.read(&mut s);
-        (c, s)
+    let c_keys = c_keygen.expand(keys.len())?;
+    let s_keys = s_keygen.expand(keys.len())?;
+    assert_eq!(&c_keys[..], &keys[..]);
+    assert_eq!(&s_keys[..], &keys[..]);
+
+    Ok(())
+}
+
+#[test]
+fn failing_handshakes() {
+    let mut rng = testing_rng();
+
+    // Set up keys.
+    let relay_secret = StaticSecret::random_from_rng(&mut rng);
+    let relay_public = PublicKey::from(&relay_secret);
+    let wrong_public = PublicKey::from([16_u8; 32]);
+    let relay_identity = RsaIdentity::from_bytes(&[12; 20]).unwrap();
+    let wrong_identity = RsaIdentity::from_bytes(&[13; 20]).unwrap();
+    let relay_ntpk = O5NtorPublicKey {
+        id: relay_identity,
+        pk: relay_public,
     };
-    assert_eq!(c_keys, s_keys);
-    assert_eq!(c_keys[..], hex!("05b858d18df21a01566c74d39a5b091b4415f103c05851e77e79b274132dc5b5")[..]);
-    // assert_eq!(c_keys[..], hex!("9c19b631fd94ed86a817e01f6c80b0743a43f5faebd39cfaa8b00fa8bcc65c3bfeaa403d91acbd68a821bf6ee8504602b094a254392a07737d5662768c7a9fb1b2814bb34780eaee6e867c773e28c212ead563e98a1cd5d5b4576f5ee61c59bde025ff2851bb19b721421694f263818e3531e43a9e4e3e2c661e2ad547d8984caa28ebecd3e4525452299be26b9185a20a90ce1eac20a91f2832d731b54502b09749b5a2a2949292f8cfcbeffb790c7790ed935a9d251e7e336148ea83b063a5618fcff674a44581585fd22077ca0e52c59a24347a38d1a1ceebddbf238541f226b8f88d0fb9c07a1bcd2ea764bbbb5dacdaf5312a14c0b9e4f06309b0333b4a")[..]);
+    let relay_ntsk = O5NtorSecretKey {
+        pk: relay_ntpk.clone(),
+        sk: relay_secret,
+    };
+
+    let wrong_ntpk1 = O5NtorPublicKey {
+        id: wrong_identity,
+        pk: relay_public,
+    };
+    let wrong_ntpk2 = O5NtorPublicKey {
+        id: relay_identity,
+        pk: wrong_public,
+    };
+
+    let resources = &Server::new_from_random(&mut rng);
+
+    // If the client uses the wrong keys, the relay should reject the
+    // handshake.
+    let mut hs_materials = CHSMaterials::new(wrong_ntpk1.clone(), "c-xxx".into());
+    let (_, handshake1) = O5NtorHandshake::client1(&hs_materials, &()).unwrap();
+    hs_materials.node_pubkey = wrong_ntpk2;
+    let (_, handshake2) = O5NtorHandshake::client1(&hs_materials, &()).unwrap();
+    hs_materials.node_pubkey = relay_ntpk;
+    let (st3, handshake3) = O5NtorHandshake::client1(&hs_materials, &()).unwrap();
+
+    let shs_materials = [SHSMaterials::new(
+        &relay_ntsk,
+        "s-yyy".into(),
+        [0u8; SEED_LENGTH],
+    )];
+    let ans1 = resources.server(&mut |_: &()| Some(()), &shs_materials, &handshake1);
+    let ans2 = resources.server(&mut |_: &()| Some(()), &shs_materials, &handshake2);
+
+    assert!(ans1.is_err());
+    assert!(ans2.is_err());
+
+    // If the relay's message is tampered with, the client will
+    // reject the handshake.
+    let (_, mut smsg) = resources
+        .server(&mut |_: &()| Some(()), &shs_materials, &handshake3)
+        .unwrap();
+    smsg[60] ^= 7;
+    let ans3 = O5NtorHandshake::client2(st3, smsg);
+    assert!(ans3.is_err());
 }
 
 #[test]
@@ -194,9 +267,8 @@ fn about_half() -> Result<()> {
     let mut not_found = 0;
     let mut not_match = 0;
     for _ in 0..1_000 {
-
         let sk = curve25519::StaticSecret::random_from_rng(&mut rng);
-        let rp: Option<curve25519::PublicRepresentative>= (&sk).into();
+        let rp: Option<curve25519::PublicRepresentative> = (&sk).into();
         let repres = match rp {
             Some(r) => r,
             None => {
@@ -206,7 +278,6 @@ fn about_half() -> Result<()> {
         };
 
         let pk = curve25519::PublicKey::from(&sk);
-
 
         let decoded_pk = curve25519::PublicKey::from(&repres);
         if hex::encode(pk) != hex::encode(decoded_pk) {
@@ -232,14 +303,13 @@ fn keypair() -> Result<()> {
         let kp = O5NtorSecretKey::generate_for_test(&mut rng);
 
         let pk = kp.pk.pk.to_bytes();
-        let repres = kp.pk.rp;
+        let repres: Option<curve25519::PublicRepresentative> = (&kp.sk).into();
 
         let pubkey = curve25519::PublicKey::from(&repres.unwrap());
         assert_eq!(hex::encode(pk), hex::encode(pubkey.to_bytes()));
     }
     Ok(())
 }
-
 
 /*
 // Benchmark Client/Server handshake.  The actual time taken that will be
